Android keyserver (#25)
* Prepending "sh" in command directive

This is necessary, because we cannot rely on the usual paths on
Android devices. However this shouldn't break compatibility with
conventional systems.

* Making key retrieval scripts Android compatible

This will exchange the commands with ones, which are compatible
with Android stock ROMs. The reference device is a Google Pixel on
unrooted stock firmware running the Dropbear fork SimpleeSSHD.

* Changing syntax slightly

This changes the syntax to be independent for the length of the
hash