Using hashes for easy identity verification (#23)
* Showing keyfiles hash for verification

If something in the setup has gone wrong, the user will notice a
difference in the hashes. This should serve as a valuable
checkpoint, before proceeding further.

* Adding hash for "nope!"

This will indicate to the user, that something in the MAC address
verification has failed and the server is responding with the
string "nope!".

* Wait for verification

* Printing keyfile's hash after creation

This way the user may take note of the hash and is able to verify
the identiy of the keyfile later more easily.

* Changing syntax slightly

This changes the syntax to be independent for the length of the
hash.