Fix documentation manual unlock 2 (#15)
* Appending/Replacing troubleshooting segment

This adds the usage of the backup initrd as a troubleshooting element.

Since the cryptsetup binaries are not included in the current
configuration, the reference has been removed for now. It will be
readded, after the configuration for including cryptsetup has been
added.

* Adding "Press F10 to continue"-reference

* Adding CRYPTSETUP=y to initramfs.conf

This is essential to unlock the device from initramfs. However this is
regarding Jessie and earlier. In Stretch this variable needs to be set
in /etc/cryptsetup-initramfs/conf-hook.

* Readding the troubleshoot segment on cryptsetup

If the script has been used, this should now work for Debian up
to Jessie.

* Adding CRYPTSETUP=y in README.md as well

This parameter should always be set.

* Rephrasing the troubleshoot section

This is to clarify, that these are two different workarounds and
do not interfere with each other.

* Better formatting

* A little bit more verbosity on CRYPTSETUP=y

It should always be set in LUKS FDE setups. It's incredibly useful.

* Revert "Adding CRYPTSETUP=y to initramfs.conf"

Jessie doesn't need this. Stretch uses a different file for that.

* Revert "Adding CRYPTSETUP=y in README.md as well"

This reverts commit 11264ef324c570ac70cb386ad55e03dbf9716856.

* Update README.md

Added manual backup creation.

* Update README.md

Updated text according to comments.