WIP Debian Stretch compatibility (#14)
* Retrieving mac address via sysfs

This should be a whole lot cleaner and durable.

* Adding CRYPTSETUP=y parameter

This will include cryptsetup in the initramfs. This is essential
to unlock the device if anything goes wrong with the automatic
unlocking.

* Verified compatibility of Stretch

It works basically just like before. ifconfig seems to have a
different output syntax and CRYPTSETUP needs to be set. These
things are fixed in the PR #14

* Changing retrieval method for MAC address

This uses sysfs to retrieve the MAC address while in initramfs.

* A little more precise wording